Create numpy.ipynb
NumPy (Numerical Python) is one of the core packages for numerical computing in Python. Pandas, Matplotlib, Statmodels, and many other Scientific libraries rely on NumPy. In Python, everything is an object, which means that even simple ints are also objects, with all the required machinery to make the object work.